Split snitch.hpp into multiple headers
