+++ include/snitch/snitch_section.hpp
@@ -1,0 +1,20 @@
+#ifndef SNITCH_SECTION_HPP
+#define SNITCH_SECTION_HPP
+
+#include "snitch/snitch_config.hpp"
+#include "snitch/snitch_test_data.hpp"
+
+namespace snitch::impl {
+struct section_entry_checker {
+    section_id  section = {};
+    test_state& state;
+    bool        entered = false;
+
+    ~section_entry_checker();
+
+    // Requires: number of sections < max_nested_sections.
+    explicit operator bool();
+};
+} // namespace snitch::impl
+
+#endif
