+++ include/snitch/snitch_type_name.hpp
@@ -1,0 +1,40 @@
+#ifndef SNITCH_TYPE_NAME_HPP
+#define SNITCH_TYPE_NAME_HPP
+
+#include "snitch/snitch_config.hpp"
+
+#include <string_view>
+
+namespace snitch::impl {
+template<typename T>
+constexpr std::string_view get_type_name() noexcept {
+#if defined(__clang__)
+    constexpr auto prefix   = std::string_view{"[T = "};
+    constexpr auto suffix   = "]";
+    constexpr auto function = std::string_view{__PRETTY_FUNCTION__};
+#elif defined(__GNUC__)
+    constexpr auto prefix   = std::string_view{"with T = "};
+    constexpr auto suffix   = "; ";
+    constexpr auto function = std::string_view{__PRETTY_FUNCTION__};
+#elif defined(_MSC_VER)
+    constexpr auto prefix   = std::string_view{"get_type_name<"};
+    constexpr auto suffix   = ">(void)";
+    constexpr auto function = std::string_view{__FUNCSIG__};
+#else
+#    error Unsupported compiler
+#endif
+
+    const auto start = function.find(prefix) + prefix.size();
+    const auto end   = function.find(suffix);
+    const auto size  = end - start;
+
+    return function.substr(start, size);
+}
+} // namespace snitch::impl
+
+namespace snitch {
+template<typename T>
+constexpr std::string_view type_name = impl::get_type_name<T>();
+}
+
+#endif
